models/main_models/rt1/requirements.txt
absl-py==2.1.0
ai2thor==3.0.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
argcomplete==3.5.0
array_record==0.5.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
aws-requests-auth==0.4.3
blinker==1.8.2
boto==2.49.0
botocore==1.35.17
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.0.0
cmake==3.25.0
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
dm-env==1.6
dm-reverb==0.14.0
dm-tree==0.1.8
dm_control==1.0.22
dmc2gymnasium @ git+https://github.com/imgeorgiev/dmc2gymnasium.git@874ec8015993222f68e6838b479c6584d10ea3d5
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
etils==1.7.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.0
gast==0.6.0
gcs-oauth2-boto-plugin==3.2
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
glide-text2im @ git+https://github.com/openai/glide-text2im.git@69b530740eb6cef69442d6180579ef5ba9ef063e
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-reauth==0.1.1
grpcio==1.66.0
gsutil==5.30
gymnasium==0.29.1
h5py==3.11.0
httplib2==0.20.4
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
immutabledict==4.2.0
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==2.14.0
kiwisolver==1.4.5
labmaze==1.0.6
lazy_loader==0.4
libclang==18.1.1
lit==15.0.7
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
mujoco==3.2.2
multidict==6.1.0
namex==0.0.8
networkx==3.3
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
portpicker==1.6.0
progressbar2==4.5.0
promise==2.3
protobuf==3.20.3
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyOpenGL==3.1.7
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-utils==3.8.2
python-xlib==0.33
pytz==2024.2
pyu2f==0.1.5
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
retry_decorator==1.1.1
rich==13.8.0
rlds==0.1.8
rsa==4.7.2
-e git+https://github.com/h2r/LaNPM-Dataset.git@5b34213bfe45de80bcbb55bdb721591778c5321b#egg=rt1_pytorch&subdirectory=models/main_models/rt1
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
sentry-sdk==2.13.0
setproctitle==1.3.3
simple_parsing==0.1.5
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-datasets==4.9.6
tensorflow-estimator==2.14.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.8.28
tokenizers==0.19.1
toml==0.10.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.5
transformers==4.44.2
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.18.1
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.14.1
yarl==1.11.1
zipp==3.20.1
